fix(ci): close the tidy-gate masking window; split the Renovate guards
Review round on #720. Both behavioural claims were settled empirically
before changing anything, because they contradicted each other.

The tidy gate could be masked by its own job (codex, P2 - confirmed)
--------------------------------------------------------------------
`go mod download` WRITES missing go.sum checksums on the runner. With the
tidy check placed after it, a commit missing only a checksum line passes:
download repairs the runner's copy, then `go mod tidy -diff` sees nothing
wrong with files that are untidy as COMMITTED. Reproduced exactly:

    missing go.sum entry, tidy-diff BEFORE download: exit=1
    same tree,            tidy-diff AFTER  download: exit=0

The step now runs before "Download dependencies", with the reasoning in the
workflow comment. (`go mod tidy -diff` itself never modifies files.)

"-diff does not fail the command" (Copilot - refuted)
-----------------------------------------------------
Verified on go1.25.4: clean tree exits 0, a stale go.sum line exits 1 -
matching the documented behaviour ("exits with a non-zero code if updates
are needed"). Declined on the thread with the exit codes.

CodeScene: module split + flattening
------------------------------------
- test_dev_stack_hygiene.py had accumulated 8 responsibilities across 35
  functions. The Renovate-vs-upstream guards (manager regexes, tool
  coverage, repo/tag matching, live upstream tags, and the 2026-08 audit
  invariants) moved to tests/unit/test_renovate_coverage.py, which owns
  exactly one concern; the hygiene module keeps the container-stack, token
  and automerge-copy guards.
- test_setup_uv_steps_pin_the_uv_version was nested four deep; the walk is
  now a comprehension helper and the test body is two flat asserts.

All 22 guards pass across both modules; non-integration suite green.
